+++ src/main/java/com/example/wikibackend/config/TenantRoutingDataSource.java
@@ -1,0 +1,12 @@
+package com.example.wikibackend.config;
+
+import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
+
+public class TenantRoutingDataSource extends AbstractRoutingDataSource {
+
+    @Override
+    protected Object determineCurrentLookupKey() {
+        // Возвращаем текущий organizationId из контекста
+        return TenantContext.getCurrentTenant();
+    }
+}
