+++ src/main/java/com/example/wikibackend/config/TenantSchemaAspect.java
@@ -1,0 +1,27 @@
+package com.example.wikibackend.config;
+
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.annotation.Before;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Component;
+
+@Aspect
+@Component
+public class TenantSchemaAspect {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public TenantSchemaAspect(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Before("execution(* com.example.wikibackend.service.*.*(..))")
+    public void setSchemaForOrganization() {
+        Long organizationId = TenantContext.getCurrentTenant();
+        if (organizationId != null) {
+            String schemaName = "alt_" + organizationId + "_data";
+            jdbcTemplate.execute("SET search_path TO " + schemaName);
+        }
+    }
+}
+
